Create an

create_bas creates an RMODFLOW bas object.

create_bas(dis = create_dis(), xsection = FALSE, chtoch = FALSE,
  free = TRUE, printtime = FALSE, showprogress = FALSE,
  stoperror = FALSE, stoper = 1, ibound = create_rmodflow_array(1, dim =
  c(dis$nrow, dis$ncol, dis$nlay)), hnoflo = -999,
  strt = create_rmodflow_array(0, dim = c(dis$nrow, dis$ncol, dis$nlay)))

Arguments

dis

RMODFLOW dis object

xsection

logical; is the model a 1-row cross-section?

chtoch

logical; should flow between adjacent constant-head cells be calculated?

free

logical; is free format used?

printtime

logical; should start, end and elapsed times be written to the global output file?

showprogress

logical; should progress information be displayed?

stoperror

logical; should the model be stopped based on budget percent discrepancy?

stoper

numeric; threshold budget percent discrepancy

ibound

3d array specifying active (1), inactive (0) or constant head (0) status of all cells

strt

3d array specifying starting heads

Value

Object of class bas

See also

read_bas, write_bas and http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?bas.htm