Create an

create_hob creates an RMODFLOW hob object

create_hob(locations, time_series, dis, hydraulic_conductivity = array(1, dim
  = c(dis$nrow, dis$ncol, dis$nlay)), iuhobsv = 0, hobdry = -888,
  noprint = FALSE, tomulth = 1, itt = 1, unique_obsnam = FALSE,
  prj = NULL)

Arguments

locations

data.frame with the name, x and y coordinates, and top and bottom of the monitoring well filter

time_series

data.frame with the name of the monitoring well filter and the observation time and head

dis

RMODFLOW dis object

hydraulic_conductivity

3d array with hydraulic conductivity values, for calculating the weights for a multilayer observation

iuhobsv

file unit for saving observation data in a file; the file unit should be included as DATA type in the name file; defaults to 0 (i.e. no observation output file will be written)

hobdry

value of the simulated equivalent for a dry cell used in the observation output file

noprint

logical; if TRUE, the input and output are not printed in the listing file

unique_obsnam

logical; should an ID number be added to obsnam for filters with observations at different times? Defaults to FALSE.

prj

RMODFLOW prj object; provide if coordinates in locations are real world coordinates

TOMULT

time offset multiplier, to convert the time unit in time_series to the MODFLOW time unit

Value

RMODFLOW hob object

See also

read.hob and write.hob