Plot a 2D section through a MODFLOW 3D array

plot.huf plots a 2D section through a MODFLOW 3D array.

# S3 method for huf
plot(huf, i = NULL, j = NULL, k = NULL, dis, bas = NULL,
  colour_palette = rev_rainbow, nlevels = 7, levels = huf$hgunam,
  type = "fill", grid = FALSE, add = FALSE, title = NULL, hed = NULL,
  l = NULL, ...)

Arguments

huf

an object of class huf

i

row number to plot

j

column number to plot

k

layer number to plot

dis

discretization file object

bas

basic file object; optional

colour_palette

a colour palette for imaging the array values

nlevels

number of levels for the colour scale; defaults to 7

levels

labels that should be used on the factor legend; huf$hgunam is used by default

type

plot type: 'fill' (default), 'factor' or 'grid'

grid

logical; should grid lines be plotted? alternatively, provide colour of the grid lines.

title

plot title

hed

hed object for only plotting the saturated part of the grid; possibly subsetted with time step number; by default, last time step is used

l

time step number for subsetting the hed object

...

parameters provided to plot.rmodflow_2d_array

mask

a 3D array with 0 or F indicating inactive cells optional; defaults to having all cells active or, if bas is provided, bas$ibound

zlim

vector of minimum and maximum value for the colour scale

Value

ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly