Analyze a MODFLOW model

This function performs local sensitivity analysis of a MODFLOW model .

rmf_analyze(
  path,
  code = "2005",
  evaluate = NULL,
  include = NULL,
  transform = NULL,
  backup = TRUE,
  restore = FALSE,
  visualize = interactive(),
  ...
)

Arguments

path
Path to the NAM file. Typically with extension .nam.

code
Name of the MODFLOW variant to use, or path to the executable.

evaluate
Vector of PVAL file parameter values to evaluate. This should be a named vector if not all parameters are provided in their order of occurrence, where the names (can be regular expressions to) match the parameter names. Parameters that are not mentioned take the value from the PVAL file. If NULL (default), no values are changed in the original PVAL file.

include
Character vector indicating which PVAL file parameters should be included. Regular expressions can be used to include multiple parameters at once. Parameters that are not mentioned are not included. If NULL (default), all parameters are included.

transform
Character vector of transformations. This should be a named vector if not all parameters are transformed, and listed in their order of occurrence, where the names (can be regular expressions to) match the parameter names. Parameters that are not mentioned are not transformed. If NULL (default), no transformations are done. Currently only "log" is supported.

backup
Logical. Should a backup (with .old suffix) of the original PVAL file be created? Defaults to TRUE.

restore
Logical. Should the original PVAL file be restored? Defaults to FALSE.

visualize
Logical. Should the results be visualized during the analysis? Defaults to TRUE in an interactive session; FALSE otherwise.

...
Optional arguments passed to rmf_execute().

Value

An rmf_analyze object, which is a list with dimensionless scaled sensitivities (dss) and composite scaled sensitivies (css).

Details

Only works with models using MODFLOW parameters and having a head predictions output file as defined in the HOB object

See also