Generic function to get cell dimensions

Generic function to get cell dimensions

Get cell dimensions from a dis object

Get cell dimensions from a huf object

rmf_cell_dimensions(...)

# S3 method for dis
rmf_cell_dimensions(dis, hed = NULL, include_volume = FALSE,
  include_faces = FALSE)

# S3 method for huf
rmf_cell_dimensions(huf, dis = NULL, hed = NULL,
  include_volume = FALSE, include_faces = FALSE)

Arguments

dis
dis object

hed
hed object, used for calculating the saturated thickness; if not specified, the regular cell thickness is returned

include_volume
logical; should the cell volumes be included?

include_faces
logical; should face areas be included?

huf
huf object

hed
hed object, used for calculating the saturated thickness; if not specified, the regular cell thickness is returned

Value

list with cell dimension 3d arrays

list with cell dimension 3d arrays