Convert a bcf to a lpf object

Convert a bcf to a lpf object

rmf_convert_bcf_to_lpf(
  bcf,
  dis,
  storagecoefficient = FALSE,
  constantcv = FALSE,
  thickstrt = FALSE,
  nocvcorrection = FALSE,
  novfc = FALSE,
  noparcheck = FALSE,
  ss_value = 1e-05,
  wetdry_value = -0.01,
  vk_bot = NULL
)

Arguments

bcf
RMODFLOW bcf object

dis
RMODFLOW dis object

storagecoefficient
logical; should STORAGECOEFFICIENT keyword be included?; defaults to FALSE

constantcv
logical; should CONSTANTCV keyword be included?; defaults to FALSE

thickstrt
logical; should THICKSTRT keyword be included?; defaults to FALSE

nocvcorrection
logical; should NOCVCORRECTION keyword be included?; defaults to FALSE

novfc
logical; should NOVFC keyword be included?; defaults to FALSE

noparcheck
logical; should NOPARCHECK keyword be included?; defaults to FALSE

ss_value
numeric; value of specific storage for layers with laycon == 1. Defaults to 1e-5. See details

wetdry_value
numeric; value of wetdry for layer with laycon == 2. Defaults to -0.01. See details.

vk_bot
2d array with the vertical hydraulic conductivity values in the bottom layer. Defaults to NULL. See details.

Value

object of class lpf

Details

Since vcont is not set for the bottom layer, the vertical conductivity in the bottom layer is assumed to be the same as hk unless a vk_bot argument is supplied. This affects the calculations of vk in the overlying layers as well.

     When confining beds are present, both vkcb and vk are unknowns. Vkcb is therefore set equal to vk and a warning is thrown.
     
     Layers with laycon = 1 in BCF should represent the upper layer of the model and have unconfined conditions. In this case, BCF needs only a specific yield value.
     When converting to LPF however, a specific storage value is also needed and can not be determined from the bcf object. The \code{ss_value} argument specifies this 
     specific storage value.
     
     Layers with laycon = 2 are converted to convertible layer types in LPF. In BCF, these layers do not have wetdry variables specified. In LPF, convertible layers can have 
     wetdry variables specified. The \code{wetdry_value} argument sets the wetdry values for layers with laycon = 2.

See also