Convert a huf to a lpf object

Convert a huf to a lpf object

rmf_convert_huf_to_lpf(
  huf,
  dis,
  mask = NULL,
  vka = ifelse(all(huf$hguvani == 0), "VK", "VANI"),
  averaging = c(HK = "arithmetic", HANI = "arithmetic", VK = "harmonic", VANI =
    "arithmetic", SS = "arithmetic", SY = "arithmetic"),
  pval = NULL,
  ...
)

Arguments

huf
RMODFLOW huf object

dis
RMODFLOW dis object

mask
masking 3d array for averaging rmf_convert_huf_to_grid, typically the ibound array, to speed up grid conversion; defaults to including all cells

vka
character indicating what variable the VKA array in the resulting lpf object represents. Possible values are 'VK' or 'VANI'. If all HGUVANI values are the same, the default vka is set correspondingly. If HGUVANI varies between hgu's, the default vka is 'VANI'.

averaging
named character vector of weighted averaging to use in rmf_convert_huf_to_grid. Possible values are 'arithmetic', 'harmonic' and 'geometric'. Names should correspond to the partyp defined in the huf object. Defaults to 'arithmetic' for every parameter type except for 'VK' ('harmonic').

pval
optional RMODFLOW pval object; used to overwrite huf parameter values in rmf_convert_huf_to_grid. Defaults to NULL.

...
arguments passed to rmf_create_lpf

Value

a RMODFLOW lpf object

Details

Huf parameters are converted to non-parameter data averaged over the dis grid using rmf_convert_huf_to_grid. The resulting lpf object will therefore not have any flow parameters. If HGUVANI varies per hgu, a correct conversion to a lpf vka array is not possible (i.e. because part of the layer might represent VANI from hgu A and another part VK from hgu B). Therefore, the user has to decide whether vka represents vk or vani in its entirety.

See also