Convert a upw to a lpf object

Convert a upw to a lpf object

rmf_convert_upw_to_lpf(
  upw,
  storagecoefficient = FALSE,
  constantcv = FALSE,
  thickstrt = FALSE,
  nocvcorrection = FALSE,
  novfc = FALSE,
  noparcheck = FALSE,
  laywet = upw$laywet,
  wetfct = 0.1,
  iwetit = 1,
  ihdwet = 0,
  wetdry = NULL
)

Arguments

upw
RMODFLOW upw object

storagecoefficient
logical; should STORAGECOEFFICIENT keyword be included?; defaults to FALSE

constantcv
logical; should CONSTANTCV keyword be included?; defaults to FALSE

thickstrt
logical; should THICKSTRT keyword be included?; defaults to FALSE

nocvcorrection
logical; should NOCVCORRECTION keyword be included?; defaults to FALSE

novfc
logical; should NOVFC keyword be included?; defaults to FALSE

noparcheck
logical; should NOPARCHECK keyword be included?; defaults to FALSE

laywet
vector of flags for each layer, indicating if wetting is active; defaults to 0 for each layer

wetfct
is a factor that is included in the calculation of the head that is initially established at a cell when it is converted from dry to wet; defaults to 0.1

iwetit
is the iteration interval for attempting to wet cells; defaults to 1

ihdwet
is a flag that determines which equation is used to define the initial head at cells that become wet; defaults to 0

wetdry
3d array with a wetting threshold and flag indicating which neighboring cells can cause a cell to become wet; defaults to NULL. If not read for a specific layer, set all values in that layer to NA.

Value

object of class lpf

Note

upw input structure is nearly identical to lpf but calculations are done differently. Differences include the addition of the iphdry value and the omission of optional keywords. Layer wetting capabilities are also not supported by upw.

See also