Create an RMODFLOW bcf object

rmf_create_bcf creates an RMODFLOW bcf object

rmf_create_bcf(
  dis,
  ibcfcb = 0,
  hdry = -888,
  iwdflg = 0,
  wetfct = 1,
  iwetit = 1,
  ihdwet = 0,
  layavg = rep(0, dis$nlay),
  laycon = rep(0, dis$nlay),
  trpy = rep(1, dis$nlay),
  sf1 = 1e-05,
  tran = 0.001,
  hy = 1e-04,
  vcont = 1e-05,
  sf2 = 0.15,
  wetdry = NULL
)

Arguments

dis
RMODFLOW dis object

ibcfcb
flag and unit number for writing cell-by-cell flow terms; defaults to 0

hdry
head assigned to cells that are converted to dry cells; defaults to -888

iwdflg
wetting capability flag; defaults to 0

wetfct
factor included in the calculation of the initial head when a cell is converted from dry to wet; defaults to 1

iwetit
iteration interval for attempting to wet cells; defaults to 1

ihdwet
flag determining which equation is used to define initial heads at cells that are converted from dry to wet; defaults to 0

layavg
numeric vector of length dis$nlay determining which method of calculating interblock transmissivity to use for each layer; defaults to 0 for 3 layers

laycon
numeric vector of length dis$nlay determining which layer type (LAYCON) to use for each layer; defaults to 0 for 3 layer

trpy
numeric vector of length dis$nlay determining the horizontal anisotropy of each layer; defaults to 1 for 3 layers

sf1
numeric 3D array of dimensions dis$nrow x dis$ncol x dis$nlay specifying the primary storage coefficient for each cell. If not read for a specific layer, set all values in that layer to NA; defaults to NULL

tran
numeric 3D array of dimensions dis$nrow x dis$ncol x dis$nlay specifying the transmissivity along rows for each cell. If not read for a specific layer, set all values in that layer to NA; defaults to 0.001 for each cell

hy
numeric 3D array of dimensions dis$nrow x dis$ncol x dis$nlay specifying the hydraulic conductivity along rows for each cell. If not read for a specific layer, set all values in that layer to NA; defaults to NULL

vcont
numeric 3D array of dimensions dis$nrow x dis$ncol x (dis$nlay-1) specifying the vertical hydraulic conductivity divided by the thickness between the node in the cell and the node in the cell below. If not read for a specific layer, set all values in that layer to NA; defaults to 1e-5 for each cell

sf2
numeric 3D array of dimensions dis$nrow x dis$ncol x dis$nlay specifying the secondary storage coefficient for each cell. If not read for a specific layer, set all values in that layer to NA; defaults to NULL

wetdry
numeric 3D array of dimensions dis$nrow x dis$ncol x dis$nlay specifying the flag and wetting threshold for each cell. If not read for a specific layer, set all values in that layer to NA; defaults to NULL

Value

RMODFLOW bcf object

See also