Create an RMODFLOW nwt object

rmf_create_nwt creates an RMODFLOW nwt object.

rmf_create_nwt(
  headtol = 0.001,
  fluxtol = 500,
  maxiterout = 100,
  thickfact = 1e-05,
  linmeth = 2,
  iprnwt = 0,
  ibotav = 0,
  options = "MODERATE",
  continue = TRUE,
  dbdtheta = 0.7,
  dbdkappa = 0.001,
  dbdgamma = 0,
  momfact = 0.1,
  backflag = 0,
  maxbackiter = 30,
  backtol = 2,
  backreduce = 0.99,
  maxitinner = 50,
  ilumethod = 2,
  levfill = 1,
  stoptol = 1e-10,
  msdr = 10,
  iacl = 2,
  norder = 1,
  level = 1,
  north = 2,
  iredsys = 0,
  rrctols = 0,
  idroptol = 1,
  epsrn = 0.001,
  hclosexmd = 1e-04,
  mxiterxmd = 50
)

Arguments

headtol
maximum head change between outer iterations for reaching convergence

fluxtol
maximum root-mean-squared flux difference between outer iterations for reaching convergence

maxiterout
maximum number of outer iterations allowed

thickfact
portion of cell thickness used for smoothly adjusting storage and conductance values to zero

linmeth
flag indicating if the GMRES (1) matrix solver or the XMD (2) matrix solver will be used

iprnwt
flag indicating if information about solver convergence is printed to the listing file

ibotav
logical; indicating whether corrections are made to groundwater head relative to the cell-bottom altitude if the cell is surrounded by dewatered cells.

options
character vector; possible values include "SPECIFIED", "SIMPLE", "MODERATE" and "COMPLEX". See details for more information.

continue
logical; should the model solve a subsequent time step after it fails to converge ? Defaults to TRUE.

dbdtheta
coefficient used to reduce the weight applied to the head change between nonlinear iterations

dbdkappa
coefficient used to increase the weight applied to the ead change between nonlinear iteraions

dbdgamma
factor used to weight the head change gor iterations n-1 and n

momfact
momentum coefficient

backflag
flag used to specify whether residual control will be used. A value of 1 indicates residual control is active, a value of 0 indicates it's inactive.

maxbackiter
is the maximum number of reductions in the head change between nonlinear iterations

backtol
proportional decrease in the root-mean-squared error of the groundwater-flow equation used to determine if residual control is required at the end of a nonlinear iteration

backreduce
reduction factor used for residual control that reduces head change between nonlinear iterations

maxitinner
read if LINMETHD = 1 & options = "SPECIFIED"; maximum number of iterations for the linear solution

ilumethod
read if LINMETHD = 1 & options = "SPECIFIED"; index for selection of the method for incomplete factorization (ILU) used as a preconditioner. See details.

levfill
read if LINMETHD = 1 & options = "SPECIFIED"; fill limit for ILUMETHOD = 1 and level of fill for ILUMETHOD = 2.

stoptol
read if LINMETHD = 1 & options = "SPECIFIED"; tolerance for convergence of the linear solver.

msdr
read if LINMETHD = 1 & options = "SPECIFIED"; number of iterations between restarts of the GMRES solver.

iacl
read if LINMETHD = 2 & options = "SPECIFIED"; flag for the acceleration method: 0 = conjugate gradient; 1 = ORTHOMIN, 2 = Bi-CGSTAB

norder
read if LINMETHD = 2 & options = "SPECIFIED"; flag for the scheme of ordering the unknowns; 0 = original ordering, 1 = RCM ordering; 2 = Minimum degree ordering

level
read if LINMETHD = 2 & options = "SPECIFIED"; level of fill for incomplte LU factorization

north
read if LINMETHD = 2 & options = "SPECIFIED"; number of orthogonalization for the ORTHOMIN acceleration scheme. Should equal 2 for the other acceleration methods

iredsys
read if LINMETHD = 2 & options = "SPECIFIED"; flag for applying reduced system preconditioning: 1 = apply; 0 = do not apply

rrctols
read if LINMETHD = 2 & options = "SPECIFIED"; residual reduction-convergence criteria

idroptol
read if LINMETHD = 2 & options = "SPECIFIED"; flag for using drop tolerance in the preconditioning

epsrn
read if LINMETHD = 2 & options = "SPECIFIED"; drop tolerance for preconditioning

hclosexmd
read if LINMETHD = 2 & options = "SPECIFIED"; head closure criteria for inner (linear) iteraions

mxiterxmd
read if LINMETHD = 2 & options = "SPECIFIED"; maximum number of iteraions for the linear solution

Value

Object of class nwt

Details

the option keyword controls the default values for input arguments dbdtheta - mxiterxmd. If option = "SPECIFIED", the user specifies the values. If option = "SIMPLE", default values will be defined by NWT that work well for nearly linear models. If option =" MODERATE", default input values will be defined by NWT that work well for moderately nonlinear models. If option = "COMPLEX", default values will be defined by NWT that work well for highly nonlinear models. See the MODFLOW-NWT manual for details on these values. Default values for dbdtheta - mxiterxmd are set to those used when option = "MODERATE".

If ILUMETHODE = 1, ILU with drop toleance and fill limit is used. If ILUMETHOD = 2, ILU(k), Order k incomplete LU factorization is used. See the MODFLOW-NWT and Kipp et al. (2008) for details.

Note

nwt must be used with the Upstream Weighting flow package. See also rmf_create_upw.

References

Kipp, K.L., Jr., Hsieh, P.A. & Charlton, S.R. 2008, Guide to the revised groundwater flow and heat transport simulator; HYDROTERM - Version 3: U.S. Geological Survey Techniques and Methods 6-A25, 160 p.

See also