Create an RMODFLOW oc object

rmf_create_oc creates an RMODFLOW oc object.

rmf_create_oc(
  dis,
  ihedfm = 0,
  chedfm = NA,
  ihedun = 666,
  iddnfm = 0,
  cddnfm = NA,
  iddnun = 667,
  cboufm = NA,
  ibouun = 668,
  compact_budget = TRUE,
  aux = TRUE,
  head_label = TRUE,
  drawdown_label = TRUE,
  ibound_label = TRUE,
  print_head = FALSE,
  print_drawdown = FALSE,
  print_budget = TRUE,
  save_head = TRUE,
  save_drawdown = FALSE,
  save_ibound = FALSE,
  save_budget = TRUE,
  incode = NULL,
  ihddfl = NULL,
  ibudfl = NULL,
  icbcfl = NULL,
  hdpr = NULL,
  ddpr = NULL,
  hdsv = NULL,
  ddsv = NULL
)

Arguments

dis
RMODFLOW dis object

ihedfm
integer code for the format in which heads will be printed; defaults to 0

chedfm
integer or character code for the format in which heads will be saved; defaults to NA (heads are saved in binary file)

ihedun
unit number on which heads will be saved; defaults to 666

iddnfm
integer code for the format in which drawdowns will be printed; defaults to 0

cddnfm
integer or character code for the format in which drawdowns will be saved; defaults to NA (drawdowns are saved in binary file)

iddnun
unit number on which drawdowns will be saved; defaults to 667

cboufm
integer or character code for the format in which ibound will be saved; defaults to NA (ibound is saved in binary file)

ibouun
unit number on which ibound will be saved; defaults to 668

compact_budget
logical; should the COMPACT BUDGET option be used

aux
logical; should the auxiliary data be included in the budget file

head_label
logical; should labels be included in the unformatted head file?

drawdown_label
logical; should labels be included in the unformatted drawdown file?

ibound_label
logical; should labels be included in the unformatted ibound file?

print_head
logical vector of length sum(dis$nstp) or logical matrix of dimensions NLAY x sum(dis$nstp); should heads be printed ? If not a matrix, all layers are treated the same.

print_drawdown
logical vector or logical matrix of dimensions NLAY x sum(dis$nstp); should drawdowns be printed ? If not a matrix, all layers are treated the same.

print_budget
logical vector; should budget be printed ?

save_head
logical vector or logical matrix of dimensions NLAY x sum(dis$nstp); should heads be saved ? If not a matrix, all layers are treated the same.

save_drawdown
logical vector or logical matrix of dimensions NLAY x sum(dis$nstp); should drawdowns be saved ? If not a matrix, all layers are treated the same.

save_ibound
logical vector or logical matrix of dimensions NLAY x sum(dis$nstp); should ibound be saved ? If not a matrix, all layers are treated the same.

save_budget
logical vector; should budget be saved ?

incode
vector of length sum(dis$nstp); used when OC is specified with numeric codes; defaults to NULL (OC is specified using words)

ihddfl
vector of length sum(dis$nstp); used when OC is specified with numeric codes; defaults to NULL (OC is specified using words)

ibudfl
vector of length sum(dis$nstp); used when OC is specified with numeric codes; defaults to NULL (OC is specified using words)

icbcfl
vector of length sum(dis$nstp); used when OC is specified with numeric codes; defaults to NULL (OC is specified using words)

hdpr
matrix of dimensions sum(dis$nstp), dis$nlay; used when OC is specified with numeric codes; defaults to NULL (OC is specified using words)

ddpr
matrix of dimensions sum(dis$nstp), dis$nlay; used when OC is specified with numeric codes; defaults to NULL (OC is specified using words)

hdsv
matrix of dimensions sum(dis$nstp), dis$nlay; used when OC is specified with numeric codes; defaults to NULL (OC is specified using words)

ddsv
matrix of dimensions sum(dis$nstp), dis$nlay; used when OC is specified with numeric codes; defaults to NULL (OC is specified using words)

Value

Object of class oc

Details

if the print_* or save_* arguments are vectors of length 1, they are repeated for every time step. The integer codes for specifying the printing formats can be found in the MODFLOW-2005 manual https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?oc.htm. A character code for specifying a saving format must be a valid FORTRAN format enclosed in parentheses and contain 20 characters or less. Alternatively, RMODFLOW allows the specification of the saving format through integer codes as used for the printing formats.

See also