Concise function for creating a 2D-array parameter.

Create a MODFLOW parameter from a 2D-array.

# S3 method for rmf_2d_array
rmf_create_parameter(
  array,
  parnam,
  parval = 1,
  kper = attr(array, "kper"),
  layer = NULL,
  partyp = NULL,
  mltnam = parnam,
  instnam = NULL
)

Arguments

array
a rmf_2d_array object

parnam
character specifying the name of the parameter

parval
numeric specifying the value of the parameter which is used to multiply values in the array. Defaults to 1.0

kper
integer vector with the stress period numbers during which the parameter is active. Specifying kper indicates that this parameter represent boundary condition information.

layer
integer vector denoting the layer indices represented by the parameter. Specifying layer indicates that this parameter represent flow package information and the partyp argument needs to be set as well.

partyp
character specifying the type of flow parameter. Allowed values are HK, HANI, VK, VANI, SS, SY and VKCB. Not used when layer is NULL.

mltnam
character vector specifying the name of the resulting multiplier array.

instnam
optional character specying the instance name of the parameter is to be time-varying; defaults to NULL

Value

a rmf_2d_array object of class rmf_parameter

Details

This function will only set the multiplier array. The user must make sure that this multiplier array is written to a separate MLT file when running a MODFLOW simulation. This function is intended to be used when no multiplier and/or zone arrays are specified for a parameter. A boundary-condition parameter is created by setting the kper argument. A flow parameter is created by setting the layer and partyp arguments.