Create an RMODFLOW upw object

rmf_create_upw creates an RMODFLOW upw object.

rmf_create_upw(
  dis,
  iupwcb = 0,
  hdry = -888,
  iphdry = TRUE,
  noparcheck = FALSE,
  laytyp = rmfi_ifelse0(dis$nlay == 1, 1, c(1, rep(0, dis$nlay - 1))),
  layavg = rep(0, dis$nlay),
  chani = rep(1, dis$nlay),
  layvka = rep(0, dis$nlay),
  parameters = NULL,
  hk = rmf_create_array(1e-04, dim = c(dis$nrow, dis$ncol, dis$nlay)),
  hani = rmfi_ifelse0(all(chani <= 0), hk * 0 + 1, NULL),
  vka = rmfi_ifelse0(all(layvka == 0), hk, NULL),
  ss = rmfi_ifelse0("TR" %in% dis$sstr, hk * 0 + 1e-05, NULL),
  sy = rmfi_ifelse0("TR" %in% dis$sstr && all(laytyp > 0), hk * 0 + 0.15, NULL),
  vkcb = NULL
)

Arguments

dis
RMODFLOW dis object

iupwcb
flag and unit number for writing cell-by-cell flow terms; defaults to 0

hdry
head assigned to cells that are converted to dry cells; defaults to -888

iphdry
logical; indicating if head will be set to hdry when it's less than 1E-4 above the cell bottom; defaults to TRUE

noparcheck
logical; should NOPARCHECK keyword be included?; defaults to FALSE

laytyp
vector of flags for each layer, specifying layer type; defaults to all confined (0) except the first layer (1)

layavg
vector of flags for each layer, specifying interblock transmissivity calculation method; defaults to 0 for each layer

chani
vector of flags or horizontal anisotropies for each layer; defaults to 1 for each layer

layvka
vector of flags for each layer, indicating whether vka is the vertical hydraulic conductivity or the ratio of horizontal to vertical; defaults to 0 for each layer

parameters
list of rmf_parameter as created by rmf_create_parameter. See details; defaults to NULL

hk
3d array with hydraulic conductivity along rows; defaults to 1. If not read for a specific layer, set all values in that layer to NA.

hani
3d array with the ratio of hydraulic conductivity along columns to that along rows; defaults to 1. If not read for a specific layer, set all values in that layer to NA.

vka
3d array with vertical hydraulic conductivity or the ratio of horizontal to vertical; defaults to hk. If not read for a specific layer, set all values in that layer to NA.

ss
3d array with specific storage; only required when there are transient stress periods; defaults to 1E-5. If not read for a specific layer, set all values in that layer to NA.

sy
3d array with specific yield; only required when there are transient stress periods; defaults to 0.15. If not read for a specific layer, set all values in that layer to NA.

vkcb
3d array with vertical hydraulic conductivity of quasi-three-dimensional confining beds; defaults to 0. If not read for a specific layer, set all values in that layer to NA.

Value

Object of class upw

Details

Flow variables are any of HK, HANI, VK, VANI, SS, SY and VKCB. A single variable can be specified either through the use of parameters or by using direct array input. When a flow variable for a specific layer is specified using parameters, all flow variables of the type must be specified by parameters. E.g. if a flow parameter defines HK for layer 1, HK must be defined for all layers using flow parameters instead of direct array input.

Note

upw input structure is nearly identical to lpf but calculations are done differently. Differences include the addition of the iphdry value and the ommision of optional keywords. Layer wetting capabilities are also not supported by upw.

upw must be used with the Newton solver. See also rmf_create_nwt.

See also