Plot a MODFLOW head file object

Plot a MODFLOW head file object

# S3 method for hed
rmf_plot(
  hed,
  dis,
  i = NULL,
  j = NULL,
  k = NULL,
  l = NULL,
  nstp = NULL,
  kper = NULL,
  kstp = NULL,
  saturated = TRUE,
  type = "fill",
  gridlines = FALSE,
  add = FALSE,
  ...
)

Arguments

hed
RMODFLOW hed object

dis
RMODFLOW dis object

i
row number to plot

j
column number to plot

k
layer number to plot

l
time step number to plot (index for 4th dimension); defaults to plotting the final time step. See details.

nstp
integer specifying the flow time step to plot. See details.

kper
integer specifying the stress-period. Use in conjunction with kstp. See details.

kstp
integer specifying the time step of kper. Use in conjunction with kper. See details.

saturated
logical indicating if the saturated grid should be used. Defaults to TRUE. See details.

...
additional parameters passed to rmf_plot.rmf_4d_array

Value

ggplot2 object or layer

Details

There are three ways to specify which time step to plot. The l argument can be specified which represents the dimension index. This can be smaller than the total number of time steps since output may not be written for all steps. The second option is to specify nstp which is the total time step number. Since output at every time step is not always written, a specified nstp value might not be available in the array. The third option is to specify both kper & kstp which specify the stress-period and corresponding time step in that stress-period. A negative kstp will plot the final time step of the stress-period.

If saturated is TRUE, the saturated grid is plotted as given by rmf_convert_dis_to_saturated_dis, which might be useful for cross-sections to show the true saturated surface.

If no output is written for the specified time step, as controlled by the Output Control file in MODFLOW, an error is thrown.