Plot a 2D section through a MODFLOW 3D array

rmf_plot.huf plots a 2D section through a MODFLOW 3D array.

# S3 method for huf
rmf_plot(
  huf,
  dis,
  i = NULL,
  j = NULL,
  k = NULL,
  hgu = NULL,
  bas = NULL,
  mask = rmf_convert_huf_to_mask(huf = huf, dis = dis, bas = bas),
  colour_palette = rmfi_rev_rainbow,
  nlevels = 7,
  levels = setNames(huf$hgunam, 1:huf$nhuf),
  type = "factor",
  gridlines = FALSE,
  crop = FALSE,
  ...
)

Arguments

huf
an object of class huf

dis
discretization file object

i
row number to plot

j
column number to plot

k
layer number to plot

hgu
character or integer of hgu to plot

bas
basic file object; optional

mask
a 3D array with 0 or F indicating inactive cells optional;

colour_palette
a colour palette for imaging continuous array values.

nlevels
number of levels for the colour scale; defaults to 7

levels
labels that should be used on the factor legend; huf$hgunam is used by default

type
plot type: 'fill', 'factor' (default) or 'grid'

gridlines
logical; should grid lines be plotted? Alternatively, provide colour of the grid lines or set to 'huf' which plots the outline of the hgu's. When hgu's are plotted, set to FALSE or 'huf'.

crop
logical; should plot be cropped by dropping NA values; defaults to FALSE

...
parameters provided to plot.rmf_2d_array

Value

ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly