Plot a 2D section through a MODFLOW 3D array

rmf_plot.rmf_3d_array plots a 2D section through a MODFLOW 3D array.

# S3 method for rmf_3d_array
rmf_plot(array, i = NULL, j = NULL, k = NULL, dis,
  bas = NULL, mask = rmfi_ifelse0(is.null(bas), array * 0 + 1, bas$ibound),
  zlim = range(array[rmfi_ifelse0(is.null(i), c(1:dim(array)[1]), i),
  rmfi_ifelse0(is.null(j), c(1:dim(array)[2]), j), rmfi_ifelse0(is.null(k),
  c(1:dim(array)[3]), k)][as.logical(mask[rmfi_ifelse0(is.null(i),
  c(1:dim(array)[1]), i), rmfi_ifelse0(is.null(j), c(1:dim(array)[2]), j),
  rmfi_ifelse0(is.null(k), c(1:dim(array)[3]), k)])], finite = TRUE),
  colour_palette = rmfi_rev_rainbow, nlevels = 7, type = "fill",
  levels = NULL, grid = FALSE, add = FALSE, title = NULL, hed = NULL,
  l = NULL, ...)

Arguments

array
an object of class rmf_3d_array

i
row number to plot

j
column number to plot

k
layer number to plot

dis
discretization file object

bas
basic file object; optional

mask
a 3D array with 0 or F indicating inactive cells optional; defaults to having all cells active or, if bas is provided, bas$ibound

zlim
vector of minimum and maximum value for the colour scale

colour_palette
a colour palette for imaging the array values

nlevels
number of levels for the colour scale; defaults to 7

type
plot type: 'fill' (default), 'factor' or 'grid'

levels
labels that should be used on the factor legend; if NULL the array factor levels are used

grid
logical; should grid lines be plotted? alternatively, provide colour of the grid lines.

title
plot title

hed
hed object for only plotting the saturated part of the grid; possibly subsetted with time step number; by default, last time step is used

l
time step number for subsetting the hed object

...
parameters provided to plot.rmf_2d_array

Value

ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly