Write a MODFLOW model

rmf_write writes all input packages in a RMODFLOW modflow object to a directory

rmf_write(
  modflow,
  file = {    
    cat("Please select nam file to overwrite or provide new filename ...\n")    
    file.choose() },
  exclude = NULL,
  suppress = FALSE,
  verbose = TRUE,
  iprn = -1
)

Arguments

modflow
RMODFLOW modflow object

file
filename of the name file to write

exclude
character vector with packages names to exclude from the simulation. Defaults to NULL

suppress
logical; remove non-supported (and thus not written) packages in the NAME file ? Defaults to FALSE

verbose
logical; should information on writing files be printed to the console ? Defaults to TRUE.

iprn
format code for printing arrays in the listing file; defaults to -1 (no printing)

Value

NULL

Details

All arrays use free-format headers INTERNAL or CONSTANT All packages will be written according to the filenames (fname) defined in the nam object. To prevent any files being overwritten, it is best to write to an empty directory.

See also