Create an RMODFLOW lpf object

rmf_create_lpf creates an RMODFLOW lpf object.

rmf_create_lpf(dis = create_dis(), ilpfcb = 0, hdry = -888, nplpf = 0,
  storagecoefficient = FALSE, constantcv = FALSE, thickstrt = FALSE,
  nocvcorrection = FALSE, novfc = FALSE, noparcheck = FALSE,
  laytyp = ifelse(dis$nlay == 1, list(1), list(c(1, rep(0, dis$nlay -
  1))))[[1]], layavg = laytyp * 0, chani = rep(1, dis$nlay),
  layvka = rep(0, dis$nlay), laywet = rep(0, dis$nlay), wetfct = 0.1,
  iwetit = 1, ihdwet = 0, parnam = NULL, partyp = NULL, parval = NULL,
  nclu = NULL, mltarr = NULL, zonarr = NULL, iz = NULL,
  hk = rmf_create_array(1e-04, dim = c(dis$nrow, dis$ncol, dis$nlay)),
  hani = rmf_create_array(1, dim = c(dis$nrow, dis$ncol, dis$nlay)),
  vka = hk, ss = rmf_create_array(1e-05, dim = c(dis$nrow, dis$ncol,
  dis$nlay)), sy = rmf_create_array(0.15, dim = c(dis$nrow, dis$ncol,
  dis$nlay)), vkcb = rmf_create_array(0, dim = c(dis$nrow, dis$ncol,
  dis$nlay)), wetdry = rmf_create_array(-0.01, dim = c(dis$nrow, dis$ncol,
  dis$nlay)))

create_lpf(...)

Arguments

dis
RMODFLOW dis object

ilpfcb
flag and unit number for writing cell-by-cell flow terms; defaults to 0

hdry
head assigned to cells that are converted to dry cells; defaults to -888

nplpf
number of lpf parameters; defaults to 0

storagecoefficient
logical; should STORAGECOEFFICIENT keyword be included?; defaults to FALSE

constantcv
logical; should CONSTANTCV keyword be included?; defaults to FALSE

thickstrt
logical; should THICKSTRT keyword be included?; defaults to FALSE

nocvcorrection
logical; should NOCVCORRECTION keyword be included?; defaults to FALSE

novfc
logical; should NOVFC keyword be included?; defaults to FALSE

noparcheck
logical; should NOPARCHECK keyword be included?; defaults to FALSE

laytyp
vector of flags for each layer, specifying layer type; defaults to all confined (0) except the first layer (1)

layavg
vector of flags for each layer, specifying interblock transmissivity calculation method; defaults to 0 for each layer

chani
vector of flags or horizontal anisotropies for each layer; defaults to 1 for each layer

layvka
vector of flags for each layer, indicating whether vka is the vertical hydraulic conductivity or the ratio of horizontal to vertical; defaults to 0 for each layer

laywet
vector of flags for each layer, indicating if wetting is active; defaults to 0 for each layer

wetfct
is a factor that is included in the calculation of the head that is initially established at a cell when it is converted from dry to wet; defaults to 0.1

iwetit
is the iteration interval for attempting to wet cells; defaults to 1

ihdwet
is a flag that determines which equation is used to define the initial head at cells that become wet; defaults to 0

parnam
vector of parameter names; names should not be more than 10 characters, are not case sensitive, and should be unique

partyp
vector of parameter types; the lpf parameter types are HK, HANI, VK, VANI, SS, SY, or VKCB

parval
vector of parameter values

nclu
vector with the number of clusters required for each parameter

mltarr
matrix of multiplier array names, with dis$nlay rows and lpf$nplpf columns; cells with non-occurring layer-parameter combinations should be NA

zonarr
matrix of zone array names, with dis$nlay rows and lpf$nplpf columns; cells with non-occurring layer-parameter combinations should be NA

iz
matrix of zone number combinations separated by spaces, with dis$nlay rows and lpf$nplpf columns; cells with non-occurring layer-parameter combinations should be NA; if zonarr is "ALL", iz should be ""

hk
3d array with hydraulic conductivity along rows; defaults to 1

hani
3d array with the ratio of hydraulic conductivity along columns to that along rows; defaults to 1

vka
3d array with vertical hydraulic conductivity or the ratio of horizontal to vertical; defaults to hk

ss
3d array with specific storage; only required when there are transient stress periods; defaults to 1E-5

sy
3d array with specific yield; only required when there are transient stress periods; defaults to 0.15

vkcb
3d array with vertical hydraulic conductivity of quasi-three-dimensional confining beds; defaults to 0

wetdry
3d array with a wetting threshold and flag indicating which neighboring cells can cause a cell to become wet; defaults to -0.01

Value

Object of class lpf

Functions

create_lpf: Deprecated function name

See also

read_lpf, write_lpf and http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?lpf.htm