Create an RMODFLOW oc object

rmf_create_oc creates an RMODFLOW oc object.

rmf_create_oc(dis = create_dis(), ihedfm = NA, chedfm = NA,
  ihedun = 666, iddnfm = NA, cddnfm = NA, iddnun = NA, cboufm = NA,
  ibouun = NA, compact_budget = TRUE, aux = FALSE, head_label = TRUE,
  drawdown_label = TRUE, ibound_label = TRUE, iperoc = rep(1:dis$nper,
  dis$nstp), itsoc = unlist(apply(data.frame(1, dis$nstp), 1, FUN =
  function(x) x[1]:x[2])), print_head = rep(FALSE, length(iperoc)),
  print_drawdown = rep(FALSE, length(iperoc)), print_budget = rep(FALSE,
  length(iperoc)), save_head = rep(TRUE, length(iperoc)),
  save_drawdown = rep(FALSE, length(iperoc)), save_ibound = rep(FALSE,
  length(iperoc)), save_budget = rep(TRUE, length(iperoc)))

create_oc(...)

Arguments

dis
dis object

ihedfm
code for the format in which heads will be printed; defaults to NA (heads are not printed)

chedfm
code for the format in which heads will be saved; defaults to NA (heads are saved in binary file)

ihedun
unit number on which heads will be saved; defaults to 666

iddnfm
code for the format in which drawdowns will be printed; defaults to NA (drawdowns are not printed)

cddnfm
code for the format in which drawdowns will be saved; defaults to NA (drawdowns are saved in binary file)

iddnun
unit number on which drawdowns will be saved; defaults to NA (drawdowns are not saved)

cboufm
code for the format in which ibound will be saved; defaults to NA (ibound is saved in binary file)

ibouun
unit number on which ibound will be saved; defaults to NA (ibound is not saved)

compact_budget
logical; should the COMPACT BUDGET option be used

aux
logical; should the auxiliary data be included in the budget file

head_label
logical; should labels be included in the unformatted head file?

drawdown_label
logical; should labels be included in the unformatted drawdown file?

ibound_label
logical; should labels be included in the unformatted ibound file?

iperoc
vector of stress period numbers

itsoc
vector of time step numbers (within the corresponding stress period)

print_head
logical vector; should heads be printed for the time step corresponding to iperoc and itsoc?

print_drawdown
logical vector; should drawdowns be printed for the time step corresponding to iperoc and itsoc?

print_budget
logical vector; should budget be printed for the time step corresponding to iperoc and itsoc?

save_head
logical vector; should heads be saved for the time step corresponding to iperoc and itsoc?

save_drawdown
logical vector; should drawdowns be saved for the time step corresponding to iperoc and itsoc?

save_ibound
logical vector; should ibound be saved for the time step corresponding to iperoc and itsoc?

save_budget
logical vector; should budget be saved for the time step corresponding to iperoc and itsoc?

Value

Object of class oc

Functions

create_oc: Deprecated function name

See also

read_oc, write_oc and http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?oc.htm