Plot a MODFLOW 2D array

rmf_plot.rmf_2d_array plots a MODFLOW 2D array.

# S3 method for rmf_2d_array
rmf_plot(array, dis, bas = NULL,
  mask = rmfi_ifelse0(is.null(bas), array * 0 + 1, {    
  warning("Using first ibound layer as mask.", call. = FALSE)     bas$ibound[, ,
  1] }), colour_palette = rmfi_rev_rainbow,
  zlim = range(array[as.logical(mask)], finite = TRUE), nlevels = 7,
  type = "fill", levels = NULL, grid = FALSE, add = FALSE,
  height_exaggeration = 100, binwidth = round(diff(zlim)/20),
  label = TRUE, prj = NULL, crs = NULL, alpha = 1, plot3d = FALSE,
  height = NULL, title = NULL)

Arguments

array
an object of class rmf_2d_array

dis
discretization file object

bas
basic file object; optional

mask
a 2D array with 0 or F indicating inactive cells; optional; defaults to having all cells active or, if bas is provided, the first layer of bas$ibound

colour_palette
a colour palette for imaging the array values

zlim
vector of minimum and maximum value for the colour scale

nlevels
number of levels for the colour scale; defaults to 7

type
plot type: 'fill' (default), 'factor', 'grid' or 'contour'

levels
labels that should be used on the factor legend; if NULL the array factor levels are used

grid
logical; should grid lines be plotted? alternatively, provide colour of the grid lines.

add
logical; if TRUE, provide ggplot2 layers instead of object, or add 3D plot to existing rgl device; defaults to FALSE

height_exaggeration
height exaggeration for 3D plot; optional

binwidth
binwidth for contour plot; defaults to 1/20 of zlim

label
logical; should labels be added to contour plot

prj
projection file object

crs
coordinate reference system for the plot

alpha
transparency value; defaults to 1

plot3d
logical; should a 3D plot be made

height
2D array for specifying the 3D plot z coordinate

title
plot title

Value

ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly